Removed TagInterface and descending classes
Currently, TagInterface was the container for resources that are marked with
that tag. Semantically, this doesn't make sense, since the tag does not
*contain* or *reference* the resources. Instead, the resource should
reference the tag. Thus TagInterface and its descending classes were removed.
Tags are now simple strings.

If you want to get the resources marked with a specific tag, use

    $locator->getByTag('tag');